server/requirements.txt
python-dotenv
google-generativeai
langchain-google-genai
langchain
langchain_community
faiss-cpu
Flask
flask-cors